% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esDesign.R
\name{AED1_SSR.SigP}
\alias{AED1_SSR.SigP}
\title{Calculate the reduced significant level in the Adaptive Enrichment Design
  (Strategy 1) with Sample Size Re-estimation Procedure}
\usage{
AED1_SSR.SigP(alpha)
}
\arguments{
\item{alpha}{The overall Type I error rate}
}
\value{
The Value of the reduced significant level
}
\description{
The \code{AED1_SSR.SigP()} is used to calculate the reduced
  significant level to protect the overall Type I error rate
}
\examples{
alpha <- 0.05
AED1_SSR.SigP(alpha = alpha)

}
\references{
\itemize{
 \item Proschan MA, Hunsberger SA. Designed extension of studies based on
         conditional power. Biometrics 1995:1315-24. <doi:10.2307/2533262>
 \item Zhao Yang, Ruitao Lin, Guosheng Yin and Ying Yuan. (2018) Sample Size
    Re-estimation in Adaptive Enrichment Trials. (In Preparation)
}
}
